Confluence() parameters should be ConfigParser defaults
This change initializes the ConfigParser in Confluence() to use as
defaults any parameters passed to __init__.  Parameters in the config
file still take precedence, but if username, password, or appid are
missing from the config file but they are passed to Confluence() the
values passed in will be used.

Added whitespace after findfile and before config = because I believe it
is in keeping with PEP8 to have the class methods set out by blank
lines.

Added debug to the named parameters to init so that the
print(config_file) line can be optionally invoked.  I believe that is a
debugging tool and not intended (or wanted) in the final.